moved descriptionSliceer to new utils fodler
