+++ js/components/researchProject.js
@@ -1,7 +1,7 @@
 import { Element } from '../../js/components/element.js';
 import { ResearchTeamMember } from '../../js/components/researchTeamMember.js';
 import { CourseResource } from '../../js/components/courseResource.js';
-import { descriptionTrim } from '../../js/descriptionSlicer.js';
+import { descriptionTrim } from '../../utils/descriptionSlicer.js';
 
 class ResearchProject extends Element {
 	constructor(
