+++ js/components/resources.js
@@ -1,6 +1,6 @@
 import { Element } from '../../js/components/element.js';
 import { ActionButton } from '../../js/components/actionButton.js';
-import { descriptionTrim } from '../../js/descriptionSlicer.js';
+import { descriptionTrim } from '../../utils/descriptionSlicer.js';
 
 let CITE_SYMBOL =
 	'<svg width="16" height="14" viewBox="0 0 16 14" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M12.5703 7.62686C12.5703 9.16849 12.5703 9.74792 11.8498 9.94165C11.3165 10.0851 10.856 10.4759 10.856 11.0282L10.856 12.8584C10.856 13.4107 11.307 13.869 11.8543 13.7953C15.0498 13.3651 16 10.7486 16 6.42653L16 1.14138C16 0.589098 15.5523 0.141382 15 0.141382L9.56874 0.141382C9.01646 0.141382 8.56874 0.589097 8.56874 1.14138L8.56874 5.42652C8.56874 5.97881 9.01646 6.42652 9.56874 6.42653L11.5703 6.42653C12.1226 6.42653 12.5703 6.87424 12.5703 7.42653L12.5703 7.62686Z" fill="#1A202C"/><path d="M2.99839 6.42653C3.55068 6.42653 3.99839 6.87424 3.99839 7.42653L3.99839 7.62686C3.99839 9.16849 3.99839 9.74792 3.27791 9.94165C2.74457 10.0851 2.28415 10.4759 2.28415 11.0282L2.28415 12.8584C2.28415 13.4107 2.7351 13.869 3.28245 13.7953C6.47777 13.3651 7.42749 10.7486 7.42749 6.42653L7.42749 1.14138C7.42749 0.589098 6.97978 0.141382 6.42749 0.141382L0.999934 0.141382C0.44765 0.141382 -6.65508e-05 0.589097 -6.65991e-05 1.14138L-6.69737e-05 5.42652C-6.7022e-05 5.97881 0.447648 6.42652 0.999933 6.42653L2.99839 6.42653Z" fill="#1A202C"/></svg>';
