+++ pages/publications/publications.js
@@ -1,7 +1,7 @@
 // imports
 import { PageRender, retrivedData } from '../../js/pageRender.js';
 import { PublicationCard } from '../../js/components/publicationCard.js';
-import { addCollapseFunction } from '../../js/descriptionSlicer.js';
+import { addCollapseFunction } from '../../utils/descriptionSlicer.js';
 import { Icons } from '../../js/components/icons.js';
 
 // Data file paths
