+++ utils/descriptionSlicer.js
